Adding truth model test scripts
truthModel.m generates a non-linear truth estimate for the states (X) with process noise, the measurements (Y) with measurement noise, and the controls (U). The sim_test function is an example implementation of the truth model.  awgn_draw is a function to add gaussian white noise to a signal with specified mean and covariance. The make_plots function simplifies the plotting process and an example implementation is also in the sim_test function.