Updates to sim_test and supporting functions.
sim_test function now includes a tuned linearized Kalman Filter along with a Monte Carlo set of simulations. The NEES statistic still looks odd at the start of the simulation, but maybe that's due to non-linear effects on the error terms? The NIS statistic looks reasonable and falls between the bounds for most time steps.