Correcting updates to keep original input structure for truthModel
I had changed the inputs and order for truthModel before I saw that you had a monte carlo generator function created. I updated my new scripts to be compatible with the original truthModel inputs. I also re-indexed the measurements to be the same as the states (with the first entry in the measurements Y being NaN)